:sparkles: Make secure logs read a key id from gpg.
The reason we need a key id for secure logs is so that we can store the
encrypted logs into a folder with the key name. This makes recovery easier
if we have multiple keys from which to choose.

Note that currently, this only works for gpg keys that are on a secure
card.